requirements.txt
streamlit==1.32.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
groq==0.5.0
requests==2.31.0
sqlparse==0.4.4
upstash-vector==0.4.0
pydantic==2.6.3
pydantic-settings==2.2.1
pytest==8.1.1
tenacity==8.2.3
pandas==2.2.0
tabulate==0.9.0
colorama==0.4.6
